requirements.txt
asgiref==3.3.4
Django==3.1.7
django-widget-tweaks==1.4.8
mysql-connector==2.2.9
mysqlclient==2.0.3
pkg-resources==0.0.0
pytz==2021.1
sqlparse==0.4.1
typing-extensions==3.7.4.3
